I separated the aside by components
