+++ src/Components/Asside/BoxClose/styled.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const StyledBocClose = styled.div`
+  display: flex;
+  justify-content: space-between;
+
+  width: 90%;
+
+  margin-top: 20px;
+
+  button {
+    background: none;
+  }
+`;
