+++ src/Components/Asside/BoxLight/styled.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const StyledBoxLight = styled.div`
+  width: 90%;
+
+  display: flex;
+  justify-content: flex-end;
+
+  button {
+    width: 40px;
+    height: 40px;
+
+    background: none;
+  }
+`;
