+++ src/Components/Asside/InputDay/index.tsx
@@ -1,0 +1,10 @@
+import { StyledDivDay } from "./styled";
+
+export const InputDay = () => {
+  return (
+    <StyledDivDay className="inputDay">
+      <h4 className="subTitleSol">Dia Terrestre</h4>
+      <input type="date" />
+    </StyledDivDay>
+  );
+};
