+++ src/Components/Asside/InputDay/styled.ts
@@ -1,0 +1,23 @@
+import styled from "styled-components";
+
+export const StyledDivDay = styled.div`
+  width: 90%;
+
+  display: flex;
+  flex-direction: column;
+
+  gap: 10px;
+
+  input {
+    border-radius: 50px;
+    height: 40px;
+
+    padding: 0 20px 0 20px;
+
+    border: none;
+
+    outline: 1px solid var(--Black);
+
+    font-size: var(--Font20);
+  }
+`;
